requirements.txt
asgiref @ file:///C:/b/abs_ae1n0p_8_w/croot/asgiref_1724072476516/work
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1730803452423/work
django-environ @ file:///home/conda/feedstock_root/build_artifacts/django-environ_1693617391635/work
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1718812538421/work
gunicorn==23.0.0
packaging==24.2
psycopg @ file:///D:/bld/psycopg-split_1727893169135/work/psycopg
psycopg-c @ file:///D:/bld/psycopg-split_1727893169135/work/psycopg-c
setuptools==75.1.0
sqlparse @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/sqlparse_1699544474746/work
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
wheel==0.44.0
